Move cartPersistance to middleware folder
