+++ Week-1/src/middleware/cart.js
@@ -1,4 +1,4 @@
-import {loadState} from '~/src/shared/cartPersistence.js'
+import {loadState} from '~/src/middleware/cartPersistence.js'
 
 export default (store) => (next) => (action) => {
   if (!action.type == 'LOAD_CART') return next(action)
